+++ src/components/Blog/Post.jsx
@@ -1,6 +1,6 @@
 import { Button } from "../index";
 
-export const Post = ({ className, title, content, date, author, btnTitle }) => {
+export const Post = ({ title, content, date, author, btnTitle }) => {
   return (
     <>
       <div className="w-96 -top-14 relative">
@@ -11,7 +11,7 @@ export const Post = ({ className, title, content, date, author, btnTitle }) => {
           className="bg-yellow-button text-sm h-8 w-30 text-black text-center items-center border mb-4 py-2 px-2"
           title={btnTitle}
         />
-        <p className="break-word text-sm font-normal mb-4">{content}</p>
+        <p className="break-word text-sm font-normal mb-4">{content}</p>cod e
         <div className="flex justify-between">
           <small className="text-gray-500">{date}</small>
           <small className="text-gray-500">{`By ${author}`}</small>
