+++ src/components/index.js
@@ -3,7 +3,7 @@
 export * from "./Blog/Draft";
 export * from "./Blog/Post";
 export * from "./Button/Button";
-export * from "./Header";
+export * from "./Header/Header";
 export * from "./Image/Image";
 export * from "./Modal";
 export * from "./MovieCardList";
